EmergencyAiVision (1)/EmergencyAiVision/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "folium>=0.14.0",
    "numpy>=2.2.5",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "requests>=2.32.3",
    "streamlit-folium>=0.15.0",
    "streamlit>=1.45.0",
    "trafilatura>=2.0.0",
    "streamlit-webrtc>=0.62.4",
    "geopy>=2.4.1",
]
